Goodie is done

+++ src/main/java/net/dohaw/blackclover/config/GrimmoireConfig.java
@@ -17,8 +17,8 @@ public class GrimmoireConfig extends Config {
         return config.getString("Display Name Color");
     }
 
-    public int getNumberSetting(SpellType spellType, String settingKey){
-        return config.getInt("Spell Settings." + spellType.getConfigKey() + "." + settingKey);
+    public double getNumberSetting(SpellType spellType, String settingKey){
+        return config.getDouble("Spell Settings." + spellType.getConfigKey() + "." + settingKey);
     }
 
     public Particle getParticle(SpellType spellType){
