+++ src/main/java/net/dohaw/blackclover/grimmoire/spell/CastSpellWrapper.java
@@ -46,7 +46,7 @@ public abstract class CastSpellWrapper extends SpellWrapper{
         super.loadSettings();
         this.regenConsumed = grimmoireConfig.getNumberSetting(KEY, "Regen Consumed");
         this.cooldown = grimmoireConfig.getNumberSetting(KEY, "Cooldown");
-        this.hotbarSlot = grimmoireConfig.getNumberSetting(KEY, "Hotbar Number");
+        this.hotbarSlot = (int) grimmoireConfig.getNumberSetting(KEY, "Hotbar Number");
     }
 
     public void deductMana(PlayerData pd){
